Updated to use standalone json config file
Also updated autocomplete entry for form fields, but didn't resolve
Including 'UserConfig.jsx'. Still need to do 'if check' on rendering' so it doesn't error out  if userConfig.jsx is not present. I think it's best to include UserConfig.jsx with the repo for now until project is closer to completion.